requirements.txt
python-telegram-bot
apscheduler
Flask
asyncio
requests
waitress